An issue simulating models with the Slab PreProcessor has been fixed (#505)
* adds IDF.copy() and IDF.saveas(inplace=True) (#254)

* Adjusts svg repr to the min/max values of the schedule (#255)

* Graceful warning when Slab or Basement program is not found

* Adds KeyBoardInterupt to IDF Thread

* catches more variations of unit name

* Adds ability to scale a schedule

* Fixes fallback limits for Schedule.plot2d() when Type is not defined

* Type can be specified in Schedule.from_values constructor

* plot2d is prettier by default

* more Typing

* Return existing object when new_object is there (#257)

* Adds ability to replace schedule values without affecting the full load hours

* more robust IDF.name property

* Keep sim files when error occurs (#276)

* updates requests requirement from ~=2.25.1 to >=2.26 (#292)

* typo

* this

* gitsubmodule as https

* p

* pp

* this

* error

* catch outfile

* logging

* Revert "logging"

This reverts commit bef3e4b7d1ae5a38bc894b636da5ebf31397bae8.

* keep

* as model

* better logging

* comments

---------

+++ archetypal/idfclass/idf.py
@@ -1485,7 +1485,8 @@ class IDF(GeomIDF):
         except KeyboardInterrupt:
             slab_thread.stop()
         finally:
-            tmp.rmtree(ignore_errors=True)
+            if not self.keep_data_err:
+                tmp.rmtree(ignore_errors=True)
             e = slab_thread.exception
             if e is not None:
                 raise e
